requirements.txt
flask
flask-jwt-extended
flask-sqlalchemy
flask-migrate
python-dotenv
qrcode[pil]
pillow